Move account suspension related coverage to concern spec
